[fix] record svgs, and toast icons
