+++ apps/mocksi-lite/content/RecordButton.tsx
@@ -7,19 +7,19 @@ const waitTime = 2000; // 2 seconds
 
 const recordingColorAndLabel = (currentStatus: AppState) => {
 	switch (currentStatus) {
+		case AppState.ANALYZING:
+			return { color: "mw-bg-orange/95", label: "Analyzing" };
 		case AppState.READYTORECORD:
 			return { color: "mw-bg-green/95", label: "Start" };
 		case AppState.RECORDING:
 			return { color: "mw-bg-crimson/95", label: "Stop" };
-		case AppState.ANALYZING:
-			return { color: "mw-bg-orange/95", label: "Analyzing" };
 		default:
 			return { color: "mw-bg-green/95", label: "Start" };
 	}
 };
 
 export const RecordButton = () => {
-	const { state, dispatch } = useContext(AppStateContext);
+	const { dispatch, state } = useContext(AppStateContext);
 
 	const handleStartRecording = () => {
 		dispatch({ event: AppEvent.START_RECORDING });
@@ -35,17 +35,15 @@ export const RecordButton = () => {
 	const { color, label } = recordingColorAndLabel(state);
 
 	if (state === AppState.READYTORECORD) {
-		// FIXME: find a better way to handle tailwindcss not seeing these classes when using syntax like `className={"mw-cursor-pointer"}`
-		const pointerClass = "mw-cursor-pointer";
 		return (
 			<div
-				className={pointerClass}
+				className="mw-cursor-pointer"
 				onClick={handleStartRecording}
 				onKeyUp={(event) => {
 					event.key === "Enter" && handleStartRecording();
 				}}
 			>
-				<img src={recordIcon} alt={"recordIcon"} />
+				<img alt={"recordIcon"} src={recordIcon} />
 			</div>
 		);
 	}
@@ -55,13 +53,13 @@ export const RecordButton = () => {
 	return (
 		<button
 			className={buttonClassNames}
-			type="button"
 			onClick={state !== AppState.ANALYZING ? handleStopRecording : undefined}
 			onKeyUp={(event) => {
 				if (event.key === "Escape" && state !== AppState.ANALYZING) {
 					handleStopRecording();
 				}
 			}}
+			type="button"
 		>
 			{state !== AppState.ANALYZING ? label : <LoadingSpinner />}
 		</button>
