+++ apps/mocksi-lite/content/Toast/index.tsx
@@ -1,24 +1,24 @@
 import type { ReactNode } from "react";
 
 interface ToastProps {
-	children: ReactNode;
-	id?: string;
-	className?: string;
-	backgroundColor?: string;
+  backgroundColor?: string;
+  children: ReactNode;
+  className?: string;
+  id?: string;
 }
 
-const Toast = ({ backgroundColor, className, id, children }: ToastProps) => {
-	const bgColor = backgroundColor ?? "mw-bg-white";
-	return (
-		<div
-			id={id}
-			className={`mw-border mw-border-solid mw-border-grey/40 mw-rounded ${bgColor} mw-flex mw-flex-row mw-items-center ${
-				className ?? ""
-			}`}
-		>
-			{children}
-		</div>
-	);
+const Toast = ({ backgroundColor, children, className, id }: ToastProps) => {
+  const bgColor = backgroundColor ?? "mw-bg-white";
+  return (
+    <div
+      className={`mw-border mw-border-solid mw-border-grey/40 mw-rounded ${bgColor} mw-flex mw-flex-row mw-items-center ${
+        className ?? ""
+      }`}
+      id={id}
+    >
+      {children}
+    </div>
+  );
 };
 
 export default Toast;
